requirements.txt
# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Audio/video processing
ffmpeg-python>=0.2.0

# GUI dependencies
PyQt6>=6.5.0

# Translation (optional)
openai>=1.0.0
requests>=2.28.0
urllib3>=1.26.0

# Optional offline translation libraries
# googletrans==4.0.0rc1
# deep-translator>=1.11.0

# Utilities
tqdm>=4.64.0
click>=8.0.0
pathlib2>=2.3.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Packaging
pyinstaller>=5.0.0
build>=0.10.0
twine>=4.0.0
